Create elasticsearch.sec
Wanted to provide some intelligence (by reformatting log entries) and sending the data off to elasticsearch. Logstash just blindly sends the verbatim log entry, so searching and filtering gets a little complicated.  With the reformatted elasticsearch document, we can easily and quickly search out data.

This is mostly just an example for people to send data to elasticsearch.